backend/requirements.txt
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6
python-docx>=1.1.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
spacy>=3.7.0,<3.8.0
redis>=5.0.0,<6.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0
pdf2image>=1.16.0
Pillow>=10.0.0
aiofiles>=24.1.0
aioredis>=2.0.1